+++ src/lib/utils/getSession.ts
@@ -6,9 +6,10 @@ import type {
 	URLFilterList,
 	compressOptions
 } from '@/lib/types';
-import { compress_options, tabAttr } from '@constants/env';
+import { compress_options, tabAttr } from '@/lib/constants/shared';
 import { compress as compressLZ } from 'lz-string';
-import { compress, getExtensionURL } from '@/lib/utils';
+import { getExtensionURL } from '@/lib/utils/extension';
+import { compress } from '@utils/compress';
 
 // Get current active tab
 export async function getCurrentTab() {
